Add BaseVertxGenerator which contains reuseable generator methods
