+++ .gitignore
@@ -7,6 +7,8 @@
 .settings
 .yardoc
 .yardopts
+.factorypath
+.checkstyle
 bin
 build
 target
